+++ src/api/routes/memberKeyImportToken.routes.spec.ts
@@ -10,10 +10,10 @@ import {
 } from '../../testUtils/stubs.js';
 import type { SuccessfulResult } from '../../utilities/result.js';
 import { mockSpy } from '../../testUtils/jest.js';
-import { HTTP_STATUS_CODES } from '../http.js';
-import type { MemberKeyImportTokenSchema } from '../schema/memberKeyImportToken.schema.js';
 import type { MemberKeyImportTokenCreationResult } from '../../memberKeyImportTokenTypes.js';
-import type { FastifyTypedInstance } from '../types/FastifyTypedInstance.js';
+import type { FastifyTypedInstance } from '../../utilities/fastify/FastifyTypedInstance.js';
+import type { MemberKeyImportTokenSchema } from '../../schemas/memberKeyImportToken.schema.js';
+import { HTTP_STATUS_CODES } from '../../utilities/http.js';
 
 const mockCreateMemberKeyImportToken = mockSpy(
   jest.fn<() => Promise<SuccessfulResult<MemberKeyImportTokenCreationResult>>>(),
