+++ src/api/routes/org.routes.spec.ts
@@ -14,13 +14,13 @@ import {
   type OrgSchemaPatch,
   type OrgSchemaMemberAccessType,
   orgSchemaMemberAccessTypes,
-} from '../schema/org.schema.js';
+} from '../../schemas/org.schema.js';
 import type { OrgCreationResult } from '../../orgTypes.js';
 import type { Result, SuccessfulResult } from '../../utilities/result.js';
 import { OrgProblemType } from '../../OrgProblemType.js';
 import { mockSpy } from '../../testUtils/jest.js';
-import { HTTP_STATUS_CODES } from '../http.js';
-import type { FastifyTypedInstance } from '../types/FastifyTypedInstance.js';
+import { HTTP_STATUS_CODES } from '../../utilities/http.js';
+import type { FastifyTypedInstance } from '../../utilities/fastify/FastifyTypedInstance.js';
 
 const mockCreateOrg = mockSpy(jest.fn<() => Promise<Result<OrgCreationResult, OrgProblemType>>>());
 const mockUpdateOrg = mockSpy(jest.fn<() => Promise<Result<undefined, OrgProblemType>>>());
