+++ src/org.ts
@@ -3,7 +3,7 @@ import isValidDomain from 'is-valid-domain';
 import type { AnyKeys } from 'mongoose';
 
 import { OrgModelSchema } from './models/Org.model.js';
-import type { OrgSchema, OrgSchemaPatch } from './services/schema/org.schema.js';
+import type { OrgSchema, OrgSchemaPatch } from './schemas/org.schema.js';
 import type { Result } from './utilities/result.js';
 import { MONGODB_DUPLICATE_INDEX_CODE, type ServiceOptions } from './serviceTypes.js';
 import {
