+++ src/orgTypes.ts
@@ -1,4 +1,4 @@
-import type { OrgSchema } from './services/schema/org.schema.js';
+import type { OrgSchema } from './schemas/org.schema.js';
 import { MemberAccessType } from './models/Org.model.js';
 
 export interface OrgCreationResult {
